Fix(control): 자동 회전 시 그래프 불연속성 문제 해결
자동 회전 모드에서 다이얼의 위상 값은 부드럽게 변하지만, 그래프 파형이 불연속적으로 보이던 문제 해결

- 원인:
  - 다이얼을 회전시키는 UI 타이머 m_autoRotateTimer의 간격 16ms과, 데이터를 샘플링하는 엔진의 타이머
m_captureTimer 간격(기본 100ms)이 서로 달라, 중간의 많은 위상 값들이 제대로 그려지지 못함.

- 해결:
  - 사용자가 설정 창에서 시간간격을 변경하면 SimulationEngine의 데이터 샘플링 간격 뿐만 아니라
MainWindow의 자동 회전 타이머 간격도 동일한 값으로 변경되도록 로직을 추가

이 변경으로 두 타이머의 속도가 항상 동기화되어, 다이얼이 움직이는 모든 단계가 그래프에 빠짐없이 반영되므로 매우 부드러운 파형이
그려진다.

+++ simulation_engine.cpp
@@ -68,11 +68,11 @@ void SimulationEngine::setAmplitude(double amplitude)
     m_amplitude = std::clamp(amplitude, config::MinVoltage, config::MaxVoltage);
 }
 
-void SimulationEngine::setPhase(int degrees)
+void SimulationEngine::setPhase(double degrees)
 {
     // 각도를 라디안으로 변환하여 저장
     // 360도 = 2 * pi 라디안
-    m_phaseRadians = static_cast<double>(degrees) * (2.0 * config::PI) / 360.0;
+    m_phaseRadians = degrees * (2.0 * config::PI) / 360.0;
 }
 
 void SimulationEngine::captureData()
